Move interactive playlist creation to music_lib_bot
