+++ src/run_tests.py
@@ -1,7 +1,7 @@
 import unittest
 from tests.test_music_util import TestMusicUtil
 from tests.test_my_music_lib import TestMyMusicLib
-from tests.test_playlist_creator import TestPlaylistCreator
+from tests.test_music_lib_bot import TestMusicLibBot
 from tests.test_spotify_client_wrapper import TestSpotifyClientWrapper
 
 if __name__ == '__main__':
